Externalized the configuration and iRODS context map parameters.
The configuration parameters that where set by a call to the init
function were grouped into a map.  Also the context map that is created
by the context-map function now contains all of the key-value pairs from
the new configuration map.  Both maps are own by the client an not the
library.  

The init function now returns the configuration map.  The functions no
longer rely on the configuration values and the iRODS context map being
defined in the namespace.  Instead they explicitly take either the
configuration or context map as arguments.  

Because the functions that rely on the context map now take it as one of
their arguments, the with-jargon macro was modified to accept a symbol
name to be bound to the context map with the scope of the macro.