Added the ability to provide back library with a mock iRODS repository.
To facilitate automated unit testing, a mock iRODS repository may be
used to back this library.  This is accomplished by providing an iRODS
proxy constructor that will be used instead of IRODSFileSystem/instance.

A missing parameter was added to the set-permissions definition.

In order to improve the readability of the with-jargon calls, the
required parameters were swapped so that the context map symbol comes
second.

Some unit tests where created to test these changes.